+++ packages/sdk/src/metadata/index.ts
@@ -4,3 +4,4 @@ export * from "./media.js";
 export * from "./metadata-attribute.js";
 export * from "./post.js";
 export * from "./profile.js";
+export * from "./verify-post-signature.js";
